+++ crates/testing/src/block_builder/random.rs
@@ -310,7 +310,7 @@ impl<TYPES: NodeType> BuilderDataSource<TYPES> for RandomBuilderSource<TYPES> {
         view_number: u64,
         sender: TYPES::SignatureKey,
         signature: &<TYPES::SignatureKey as SignatureKey>::PureAssembledSignatureType,
-        _num_nodes: u64,
+        _num_nodes: usize,
     ) -> Result<AvailableBlockData<TYPES>, BuildError> {
         self.claim_block(block_hash, view_number, sender, signature)
             .await
